+++ cli/cmd/root.go
@@ -5,12 +5,12 @@ import (
 	"os"
 	"time"
 
-	"github.com/google/uuid"
-	"github.com/rs/zerolog"
-
 	analytics "github.com/cloudquery/cloudquery/cli/internal/analytics"
 	"github.com/cloudquery/cloudquery/cli/internal/enum"
 	"github.com/cloudquery/cloudquery/cli/internal/env"
+	"github.com/cloudquery/cloudquery/cli/internal/uuid"
+	guuid "github.com/google/uuid"
+	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 	"github.com/thoas/go-funk"
@@ -42,6 +42,11 @@ func NewCmdRoot() *cobra.Command {
 	noLogFile := false
 	logFileName := "cloudquery.log"
 	sentryDsn := sentryDsnDefault
+	var err error
+	if invocationUUID.UUID, err = guuid.NewRandom(); err != nil {
+		fmt.Fprintf(os.Stderr, "failed to generate invocation uuid: %v", err)
+		os.Exit(1)
+	}
 
 	// support legacy telemetry environment variable,
 	// but the newer CQ_TELEMETRY_LEVEL environment variable takes precedence
@@ -50,9 +55,9 @@ func NewCmdRoot() *cobra.Command {
 	if legacyTelemetry != "" {
 		defaultTelemetryValue = "none"
 	}
-	err := telemetryLevel.Set(env.GetEnvOrDefault("CQ_TELEMETRY_LEVEL", defaultTelemetryValue))
+	err = telemetryLevel.Set(env.GetEnvOrDefault("CQ_TELEMETRY_LEVEL", defaultTelemetryValue))
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "failed to set telemetry level: "+err.Error())
+		fmt.Fprintf(os.Stderr, "failed to set telemetry level: %v", err)
 		os.Exit(1)
 	}
 
@@ -70,11 +75,6 @@ func NewCmdRoot() *cobra.Command {
 			// PersistentPreRunE runs after argument parsing, so errors during parsing will result in printing the help
 			cmd.SilenceUsage = true
 			var err error
-
-			invocationUUID, err = uuid.NewRandom()
-			if err != nil {
-				return fmt.Errorf("failed to generate invocation uuid: %w", err)
-			}
 
 			if logFile, err = initLogging(noLogFile, logLevel, logFormat, logConsole, logFileName); err != nil {
 				return err
@@ -133,6 +133,8 @@ func NewCmdRoot() *cobra.Command {
 	f := cmd.PersistentFlags().VarPF(telemetryLevel, "telemetry-level", "", "Telemetry level (none, errors, stats, all)")
 	f.DefValue = "all"
 
+	iUUID := cmd.PersistentFlags().VarPF(&invocationUUID, "invocation-id", "", "useful for when using Open Telemetry integration for tracing and logging to be able to correlate logs and traces through many services")
+	iUUID.DefValue = "<NEW-RANDOM-UUID>"
 	cmd.SetHelpCommand(&cobra.Command{Hidden: true})
 
 	pluginCmd := &cobra.Command{