+++ plugins/source/k8s/views/views.go
@@ -1,6 +1,0 @@
-package views
-
-import _ "embed"
-
-//go:embed resource.sql
-var ResourceView string